5758 - Display error tooltip
